requirements.txt
python 3.7

tensorflow
tensorboard
Keras
opencv-python
numpy
Pillow
tqdm

Formatter:  pip3.7 install black
